requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==3.2.2
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
ecdsa==0.19.2
fastapi==0.135.3
greenlet==3.4.0
h11==0.16.0
idna==3.11
Mako==1.3.11
MarkupSafe==3.0.3
passlib==1.7.4
pyasn1==0.6.3
pycparser==3.0
pydantic==2.13.0
pydantic_core==2.46.0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.26
requests==2.33.1
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.49
SQLAlchemy-Utils==0.42.1
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
